requirements.txt
psycop2-binary
flask
python-dotenv